test: add a directive function to build the tests
