+++ tests/test_execute/test_basic_singlehtml_.html
@@ -1,0 +1,18 @@
+<div class="jupyter_cell jupyter_container docutils container">
+ <div class="cell_input code_cell docutils container">
+  <div class="highlight-ipython3 notranslate">
+   <div class="highlight">
+    <pre><span></span><span class="mi">2</span> <span class="o">+</span> <span class="mi">2</span>
+</pre>
+   </div>
+  </div>
+ </div>
+ <div class="cell_output docutils container">
+  <div class="output text_plain highlight-none notranslate">
+   <div class="highlight">
+    <pre><span></span>4
+</pre>
+   </div>
+  </div>
+ </div>
+</div>
